@@ -1,8 +1,8 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-  ext.kotlin_version = '1.2.21'
-  ext.ktx_version = '0.1'
+  ext.kotlin_version = '1.3.21'
+  ext.ktx_version = '1.0.1'
   ext.multidexVersion = '1.0.1'
   ext.anko_version = '0.10.4'
   repositories {
@@ -10,10 +10,10 @@ buildscript {
     jcenter()
   }
   dependencies {
-    classpath 'com.android.tools.build:gradle:3.0.1'
+    classpath 'com.android.tools.build:gradle:3.4.0'
     classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-    classpath 'com.google.gms:google-services:3.2.0'
-    classpath "com.jakewharton:butterknife-gradle-plugin:8.8.1"
+    classpath 'com.google.gms:google-services:4.2.0'
+    classpath "com.jakewharton:butterknife-gradle-plugin:10.0.0"
 
     // NOTE: Do not place your application dependencies here; they belong
     // in the individual module build.gradle files
@@ -35,11 +35,11 @@ task clean(type: Delete) {
 ext {
 
   // SDK
-  compileSdkVersion = 27
-  buildToolsVersion = "27.0.2"
+  compileSdkVersion = 28
+  buildToolsVersion = "28.0.3"
 
   // App dependencies
-  androidSupportVersion = '27.0.2'
+  androidSupportVersion = '28.0.3'
   leakcanaryVersion = '1.5'
   rxVersion = '1.2.1'
   okHttpVersion = '3.9.0'
@@ -49,7 +49,7 @@ ext {
   googleAutoVersion = '1.0-beta4'
   guavaVersion = '20.0'
   jwtVersion = '0.7.0'
-  butterknifeVersion = '8.8.1'
+  butterknifeVersion = '10.0.0'
   glideVersion = '3.7.0'
   datePickerVersion = '3.4.1'
   jodaVersion = '2.9.4'
