@@ -1,6 +1,6 @@
 package com.tickr.tickr.dagger.modules.activities
 
-import android.support.v7.app.AlertDialog
+import androidx.appcompat.app.AlertDialog
 import com.tickr.tickr.api.managers.ApiManager
 import com.tickr.tickr.dagger.scopes.UserScope
 import com.tickr.tickr.ui.activities.detailednews.DetailedNewsActivity
