@@ -2,7 +2,7 @@ package com.tickr.tickr.ui
 
 import android.app.Activity
 import android.content.DialogInterface
-import android.support.v7.app.AlertDialog
+import androidx.appcompat.app.AlertDialog
 import com.tickr.tickr.R
 
 /**
