@@ -1,12 +1,12 @@
 package com.tickr.tickr.ui.activities.newslist
 
 import android.app.Activity
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.tickr.tickr.R
 import com.tickr.tickr.models.Article
@@ -20,20 +20,24 @@ import com.tickr.tickr.ui.utils.OnBindViewListener
 class PlatformListAdapter(var articles: ArrayList<Article>,
     var platformListActivity: PlatformListActivity,
     var presenter: PlatformListPresenter) : RecyclerView.Adapter<PlatformListAdapter.ViewHolder>() {
-  override fun onCreateViewHolder(parent: ViewGroup?, viewType: Int): ViewHolder {
-    return ViewHolder(LayoutInflater.from(platformListActivity).inflate(R.layout.item_platform_news, parent, false))
+
+  override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+    return ViewHolder(
+        LayoutInflater.from(platformListActivity).inflate(R.layout.item_platform_news, parent,
+            false))
   }
+
+  override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+    holder?.onBind(articles[position], platformListActivity, presenter)
+  }
+
 
   override fun getItemCount(): Int {
     return articles.size
   }
 
-  override fun onBindViewHolder(holder: ViewHolder?, position: Int) {
-    holder?.onBind(articles[position], platformListActivity, presenter)
-  }
 
-
-  class ViewHolder(itemView: View?) : RecyclerView.ViewHolder(itemView), OnBindViewListener {
+  class ViewHolder(itemView: View?) : RecyclerView.ViewHolder(itemView!!), OnBindViewListener {
 
     var ivPlatformNews = itemView?.findViewById<ImageView>(R.id.ivPlatformNews)
     var tvTitle = itemView?.findViewById<TextView>(R.id.tvTitle)
@@ -47,7 +51,9 @@ class PlatformListAdapter(var articles: ArrayList<Article>,
       tvDescription?.text = article.description
       tvAuthor?.text = article.author
 
-      itemView.setOnClickListener({ (presenter as PlatformListPresenter).onSingleItemClick(article) })
+      itemView.setOnClickListener {
+        (presenter as PlatformListPresenter).onSingleItemClick(article)
+      }
     }
 
   }
