@@ -1,9 +1,3 @@
-/*
- *
- *  *  Copyright (c) 2017. GreatFeat Services, Inc.
- *
- */
-
 package com.tickr.tickr.ui.utils
 
 import android.content.Context
@@ -12,11 +6,12 @@ import android.graphics.drawable.BitmapDrawable
 import android.graphics.drawable.ColorDrawable
 import android.graphics.drawable.Drawable
 import android.net.Uri
-import android.support.annotation.ColorInt
-import android.support.annotation.ColorRes
-import android.support.annotation.DrawableRes
 import android.util.AttributeSet
 import android.widget.ImageView
+import androidx.annotation.ColorInt
+import androidx.annotation.ColorRes
+import androidx.annotation.DrawableRes
+import androidx.appcompat.widget.AppCompatImageView
 import com.tickr.tickr.R
 
 /**
@@ -25,7 +20,7 @@ import com.tickr.tickr.R
  * @author edwardbryan.abergas@gmail.com
  */
 
-class CustomCircleImageView : android.support.v7.widget.AppCompatImageView {
+class CustomCircleImageView : AppCompatImageView {
 
   private val mDrawableRect = RectF()
   private val mBorderRect = RectF()
